# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.CategoryRequest:
    properties:
      name:
        example: Electronics
        type: string
    type: object
...
  handler.CreateOrderRequest:
    properties:
      district:
        example: Chatuchak
        type: string
      logistic:
        example: Kerry Express
        type: string
      logistic_branch:
        example: Branch 001
        type: string
      phone:
        example: "0812345678"
        type: string
      province:
        example: Bangkok
        type: string
      receiver_name:
        example: John Doe
        type: string
    type: object
...
  handler.UpdateOrderStatusRequest:
    properties:
      status:
        example: shipped
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample e-commerce server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Golang E-Commerce API
  version: "1.0"
paths:
  /admin/categories/create:
    post:
      consumes:
      - application/json
      description: Create a new product category
      parameters:
      - description: Category details (name)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new category (Admin)
      tags:
      - admin
      - categories
  /admin/categories/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a category (Admin)
      tags:
      - admin
      - categories
  /admin/categories/update/{id}:
    put:
      consumes:
      - application/json
      description: Update the name of an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category details (name)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a category (Admin)
      tags:
      - admin
      - categories
  /admin/orders:
    get:
      consumes:
      - application/json
      description: Retrieve all orders in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all orders (Admin)
      tags:
      - admin
      - orders
  /admin/orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of an existing order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Status Update (status)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update order status (Admin)
      tags:
      - admin
      - orders
...
  /categories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all categories
      tags:
      - categories
  /categories/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a product category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get category by ID
      tags:
      - categories
...
  /user/orders:
    get:
      consumes:
      - application/json
      description: Retrieve all orders placed by the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user's orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create an order from the user's current shopping cart
      parameters:
      - description: Order details (receiver_name, phone, province, district, logistic,
          logistic_branch)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - orders
  /user/orders/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"